requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
alembic==1.13.0

# Authentication & Security
bcrypt==4.1.2
python-jose==3.3.0
passlib==1.7.4

# Validation & Sanitization
email-validator==2.1.0
bleach==6.1.0
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
coverage==7.3.4

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==1.25.2

# Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Profiling
memory-profiler==0.61.0
py-spy==0.3.14

# Part II - Caching
redis==5.0.1
hiredis==2.3.2

# Part II - Messaging
pika==1.3.2
celery==5.3.4

# Part II - Monitoring
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0

# HTTP Requests (Inter-service communication)
requests==2.31.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
gunicorn==21.2.0

# Rate Limiting
Flask-Limiter==3.5.0

# Logging
python-json-logger==2.0.7
